refactor: splitting the logic from middleware
